a fix code style

+++ modules/task_1/zotov_m_integration_Monte_Carlo_method/integration_Monte_Carlo_method.cpp
@@ -6,12 +6,10 @@
 #include "../../../modules/task_1/zotov_m_integration_Monte_Carlo_method/integration_Monte_Carlo_method.h"
 
 
-double f1(double x)
-{
+double f1(double x) {
     return pow(x, 2);
 }
-double f2(double x)
-{
+double f2(double x) {
     return x / 2 + 1;
 }
 double f3(double x) {
@@ -24,8 +22,7 @@ double f5(double x) {
     return pow(x, 2) / 3 + 1;
 }
 
-double integralMonteCarlo(double a, double b, int n, double(*f)(double))
-{
+double integralMonteCarlo(double a, double b, int n, double(*f)(double)) {
     std::mt19937 generate;
     generate.seed(time(NULL));
     std::uniform_real_distribution<> uid(0, RAND_MAX);
@@ -33,7 +30,7 @@ double integralMonteCarlo(double a, double b, int n, double(*f)(double))
     double x;
     double y = 0.0;
 
-    for (size_t i = 0; i < n; i++) {
+    for (int i = 0; i < n; i++) {
         x = (uid(generate)/RAND_MAX) * (static_cast<double>(b - a) + a);
         y += f(x);
     }
@@ -41,11 +38,11 @@ double integralMonteCarlo(double a, double b, int n, double(*f)(double))
     return res;
 }
 
-double integralParallel(double a, double b, int n, double(*f)(double))
-{
+double integralParallel(double a, double b, int n, double(*f)(double)) {
     std::mt19937 generate;
     generate.seed(time(NULL));
     std::uniform_real_distribution<> uid(0, RAND_MAX);
+
     int ProcRank, ProcNum;
     MPI_Comm_size(MPI_COMM_WORLD, &ProcNum);
     MPI_Comm_rank(MPI_COMM_WORLD, &ProcRank);
@@ -53,14 +50,11 @@ double integralParallel(double a, double b, int n, double(*f)(double))
     double x;
     double y = 0.0;
 
-    for (size_t i = 0; i < n; i += ProcNum) {
+    for (int i = 0; i < n; i += ProcNum) {
         x = (uid(generate) / RAND_MAX) * (static_cast<double>(b - a) + a);
         y += f(x);
     }
     myRes = (static_cast<double>(b - a) + a) * y / n;
-
     MPI_Reduce(&myRes, &res, 1, MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
-    
     return res;
-
 }
